+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Taylor
 
-## Unreleased
+## V0.3.8
 
 - Fix documentation typo
 - Allow memory growth on the web builds